Add files via upload
The checker searches for all the words that are in the small text file inside the larger big text file and then counts how many times they appear in the big text file. If a word that is in the small file does not appear in the big text, it automatically leaves it out since the count is equal to zero. Due to the Github size restrictions, a smaller file size is uploaded but the same code can be used for even larger file sizes.